Move the person model classes to a more suitable location
